initialization of directories structure moved to root PersistentPreRun
